Refactor fstar handlers to take in fresh server arguments
This change is meant to fix the bug Gabriel pointed out where the
handlers could operate on stale configurations if they close over
configuration settings. I instead opted here to refactor the handlers to
not be factory functions unless they really need to (e.g. to close over
the private buffer for `handleFStarResponseForDocument`). The other
functions can simply take the server parameter as an argument when
called. The server should be passed by reference so it is not expensive
to pass it with each function call.