Refactor F* IDE connection into separate class
In addition to introducing the new FStarConnection class, this commit
introduces a custom Result type as well as a couple of custom Error
types that make it possible for the code that wraps the F* interactions
to be fully decoupled from the LSP connection (e.g. errors are returned
using Result and Error objects instead of sent over a ClientConnection).
Additionally, this commit moves the message buffering logic to the
FStarConnection class and exposes a 'message' event type on top of that
buffering logic.

Further refactoring is needed to get rid of fstar_handlers.ts entirely.
Further refactoring should also likely be done to combine
FStarConnection and FStar.