Move F* message parsing logic into new events on FStarConnection
The key idea with this commit is that only the classes that wrap F*
should worry about parsing messages from F*. Consumers of those messages
(e.g. the LSP server) should be able to interact with those messages
using a higher-level interface. In this case I decided to emit custom
events for each message that was previously being parsed.

This commit also introduces parsers and interface checkers for each of
the possible F* messages. Most of these methods are not used and could
be removed to reduce the codebase that needs to be maintained, but
ideally there should be a clean way to parse the messages
and tell what kind of message they are.