WIP - Add async F* connection API
This commit adds the FStarConnection class which handles communication
with F*. The class has async methods that can be called to send requests
to F*. In the case that a response is expected, the response can be
accessed via the returned promise.

This commit is titled WIP since there are many outstanding TODOs and
changes that need to be made depending on the exact behavior of the F*
IDE protocol. For example, the proof-status and IdeDiagnostic messages
sent by F* are currently not handled anywhere.