Return full response messages instead of response/content fields
full-buffer queries receive many different kinds of responses they need
to handle. Returning the full response message instead of either the
response or content fields allows for better differentiating between the
kind of response. Furthermore, there is a status field on the responses
that was previously being stripped away. The status field is not
currently used, but returning the full response message allows for
different behavior based on the status field in the future.