Close pipe before killing F*
The F* IDE mode will not die when a SIGINT is received. We need to close
the pipe to signal no more data will be written for it to gracefully
exit.

This is easily reproducible (before this patch) by opening a file,
starting F*, and then closing the file. The F* process remains alive,
eating up RAM.